Split `loadAnnotationsService` out of `annotationsService`
The `load` and `searchAndLoad` methods previously within
`annotationsService` didn't have a lot of functional or dependency
overlap with the other methods in `annotationsService`. Simplify
tests and make module responsibilities clearer by splitting
this into two, introducing a new `loadAnnotationsService`.

Update the `sidebar-content` controller to use
`loadAnnotationsService`.
