+++ src/sidebar/services/annotations.js
@@ -1,4 +1,8 @@
-import SearchClient from '../search-client';
+/**
+ * A service for creating, manipulating and persisting annotations and their
+ * application-store representations. Interacts with API services as needed.
+ */
+
 import * as metadata from '../util/annotation-metadata';
 import {
   defaultPermissions,
@@ -9,14 +13,26 @@ import { generateHexString } from '../util/random';
 import uiConstants from '../ui-constants';
 
 // @ngInject
-export default function annotationsService(
-  annotationMapper,
-  api,
-  store,
-  streamer,
-  streamFilter
-) {
-  let searchClient = null;
+export default function annotationsService(api, store) {
+  /**
+   * Apply changes for the given `annotation` from its draft in the store (if
+   * any) and return a new object with those changes integrated.
+   */
+  function applyDraftChanges(annotation) {
+    const changes = {};
+    const draft = store.getDraft(annotation);
+
+    if (draft) {
+      changes.tags = draft.tags;
+      changes.text = draft.text;
+      changes.permissions = draft.isPrivate
+        ? privatePermissions(annotation.user)
+        : sharedPermissions(annotation.user, annotation.group);
+    }
+
+    // Integrate changes from draft into object to be persisted
+    return { ...annotation, ...changes };
+  }
 
   /**
    * Extend new annotation objects with defaults and permissions.
@@ -101,75 +117,6 @@ export default function annotationsService(
   }
 
   /**
-   * Load annotations for all URIs and groupId.
-   *
-   * @param {string[]} uris
-   * @param {string} groupId
-   */
-  function load(uris, groupId) {
-    annotationMapper.unloadAnnotations(store.savedAnnotations());
-
-    // Cancel previously running search client.
-    if (searchClient) {
-      searchClient.cancel();
-    }
-
-    if (uris.length > 0) {
-      searchAndLoad(uris, groupId);
-
-      streamFilter.resetFilter().addClause('/uri', 'one_of', uris);
-      streamer.setConfig('filter', { filter: streamFilter.getFilter() });
-    }
-  }
-
-  function searchAndLoad(uris, groupId) {
-    searchClient = new SearchClient(api.search, {
-      incremental: true,
-    });
-    searchClient.on('results', results => {
-      if (results.length) {
-        annotationMapper.loadAnnotations(results);
-      }
-    });
-    searchClient.on('error', error => {
-      console.error(error);
-    });
-    searchClient.on('end', () => {
-      // Remove client as it's no longer active.
-      searchClient = null;
-
-      store.frames().forEach(function(frame) {
-        if (0 <= uris.indexOf(frame.uri)) {
-          store.updateFrameAnnotationFetchStatus(frame.uri, true);
-        }
-      });
-      store.annotationFetchFinished();
-    });
-    store.annotationFetchStarted();
-    searchClient.get({ uri: uris, group: groupId });
-  }
-
-  /**
-   * Apply changes for the given `annotation` from its draft in the store (if
-   * any) and return a new object with those changes integrated.
-   */
-  function applyDraftChanges(annotation) {
-    const changes = {};
-    const draft = store.getDraft(annotation);
-
-    if (draft) {
-      changes.tags = draft.tags;
-      changes.text = draft.text;
-      changes.permissions = draft.isPrivate
-        ? privatePermissions(annotation.user)
-        : sharedPermissions(annotation.user, annotation.group);
-    }
-
-    // Integrate changes from draft into object to be persisted
-    return { ...annotation, ...changes };
-  }
-
-  /**
    * Create a reply to `annotation` by the user `userid` and add to the store.
    *
    * @param {Object} annotation
@@ -225,7 +172,6 @@ export default function annotationsService(
 
   return {
     create,
-    load,
     reply,
     save,
   };
